server/requirements.txt
aiohttp
python-socketio
psutil